[ros] change XmlRpc namespace to avoid conflict
YARP and ROS both use the XmlRpc++ library.  When YARP is
compiled statically, these libraries can conflict if the
user is also linking ROS.  (This is not a problem when
YARP is compiled as a shared library, where the
XmlRpc-related symbols are not exposed).  To support
static linking, the XmlRpc namespace in YARP is renamed
to YarpXmlRpc.  See #167